Remove conditional from Game#pick_square
- Previously added conditional to restrict user input to (1..9) as
I had concerns about a user input of 'X' which would get a match in the
Board#get_square method and would allow a square to be replayed.
Now realized that the #to_i method applied to the user input defaults
strings to 0 which can never match any square on the board.
Decided not to amend the commit to keep it as a lesson